requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Detection models
ultralytics>=8.0.0  # For RT-DETR
retinaface-pytorch>=0.0.1

# Tracking
filterpy>=1.4.5
scikit-learn>=1.3.0
scipy>=1.11.0

# Video processing
ffmpeg-python>=0.2.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0

# Optional: GPU optimization
onnxruntime-gpu>=1.16.0  # For ONNX inference
#tensorrt>=8.6.0  # For TensorRT optimization (if available)